chore: remove cache (#7)
I don't think this caching is working because we are using latest and not a specific version which means there is no way of knowing when to invalidate the cache. 

There is a way to implement it by checking what is the latest version but this is more complex and not needed right now.